day4/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
lxml==6.0.2
PyPDF2==3.0.1
python-docx==1.2.0
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
fastapi
uvicorn
python-multipart
python-dotenv
